src/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
black==22.1.0
boto3==1.21.3
botocore==1.24.3
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
docutils==0.17.1
esbonio==0.11.2
et-xmlfile==1.1.0
exchange-calendars==3.6.1
finviz==1.4.3
frozenlist==1.3.0
furo==2022.6.4.1
h11==0.13.0
idna==3.3
imagesize==1.3.0
imap-tools==0.54.0
importlib-metadata==4.10.1
investpy==1.0.8
Jinja2==3.0.3
jmespath==0.10.0
korean-lunar-calendar==0.2.1
lxml==4.7.1
MarkupSafe==2.0.1
multidict==6.0.2
multitasking==0.0.10
mypy-extensions==0.4.3
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.0
pandas-datareader==0.10.0
pandas-market-calendars==3.4
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.8.2
pygls==0.11.3
Pygments==2.11.2
pyluach==1.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyspellchecker==0.6.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.5
python-magic==0.4.25
pytz==2021.3
requests==2.27.1
s3transfer==0.5.1
selenium==4.1.0
selenium-stealth==1.0.6
sendgrid==6.9.5
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.4.0
sphinx-autodoc-typehints==1.16.0
sphinx-basic-ng==0.0.1a11
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starkbank-ecdsa==2.0.3
tenacity==7.0.0
tomli==2.0.1
toolz==0.11.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
user-agent==0.1.10
webdriver-manager==3.5.2
wsproto==1.0.0
yahoo-earnings-calendar==0.6.0
yarl==1.7.2
yfinance==0.1.70
zipp==3.7.0
